+++ CHANGELOG.md
@@ -213,4 +213,4 @@
 - Remove `gerweck-utils-twitter`
   - This is supported in Twitter's bijection library
 - Switch from `org.spire-math:spire:0.13.0` to `org.typelevel:spire:0.16.2`
-
+- Add new `BuildingFlow` in Akka utils